eqLaws: Use "Reflexivity", "Symmetry" and "Transitivity"
instead of "Reflexive", "Symmetric" and "Transitive".

This is consistent with how `semigroupLaws` and `monoidLaws` use
"Associativity" and "Identity".

It is also consistent with the terminology used in the Haddocks of
`base`:
https://hackage.haskell.org/package/base-4.19.0.0/docs/Data-Eq.html#t:Eq

My practical motivation for this change is that it will make the output
for `hspec` consistent and grammatically correct. (see
https://github.com/hedgehogqa/haskell-hedgehog-classes/issues/41#issuecomment-1807931945)